Really basic splash screen render.
* New VTAG.Graphics project handles loading splash screens from bitmap objects.

This adds a dependency on System.Drawing. I create a Test project in anticipation
of future need (realistically, Console would need to be abstracted to make the rendering
testable in any reasonably automated fashion.) The test harness project demonstrates
loading a couple of demo images and current behavior when one can't be loaded.